Organised nicely into respective packages
All classes havee been organized into suitable java packages
-> command related files are in bob.command
-> task related files are in bob.task
-> ui related files are in bob.ui
-> storage related files are in bob.storage
-> exception files are in bob.exception

Also Fixed testing files
