+++ src/main/java/bob/command/ByeCommand.java
@@ -1,4 +1,10 @@
-public class ByeCommand implements Command{
+package bob.command;
+
+import bob.storage.Storage;
+import bob.task.TaskList;
+import bob.ui.Ui;
+
+public class ByeCommand implements Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         storage.save(tasks);
