+++ src/main/java/bob/exception/BobDateTimeException.java
@@ -1,3 +1,5 @@
+package bob.exception;
+
 public class BobDateTimeException extends RuntimeException {
     public BobDateTimeException(String message) {
         super(" What the Bob!!![BobDateTimeException]\n " + message);
