+++ src/main/java/bob/exception/BobException.java
@@ -1,3 +1,5 @@
+package bob.exception;
+
 public class BobException extends Exception{
     public BobException(String message){
         super(" What the BOB!!![BobException]\n"+message);
