+++ src/main/java/bob/task/EventTask.java
@@ -1,4 +1,9 @@
-import java.time.LocalDate;
+package bob.task;
+
+import bob.command.CommandFormat;
+import bob.exception.BobDateTimeException;
+import bob.exception.BobInvalidFormatException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
