+++ src/main/java/bob/task/Task.java
@@ -1,3 +1,8 @@
+package bob.task;
+
+import bob.exception.BobDateTimeException;
+import bob.exception.BobInvalidFormatException;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
