+++ src/main/java/bob/task/TaskType.java
@@ -1,3 +1,5 @@
+package bob.task;
+
 public enum TaskType {
     TODO("T"),
     DEADLINE("D"),
