+++ src/main/java/bob/task/ToDoTask.java
@@ -1,3 +1,5 @@
+package bob.task;
+
 public class ToDoTask extends Task{
     public ToDoTask(String description){
         super(description, TaskType.TODO);
