+++ text-ui-test/EXPECTED.TXT
@@ -12,24 +12,23 @@ ____________________________________________________________
 ____________________________________________________________
  Aite. I've bobbed it into the list:
    [T][ ] hi
- Bobbing heck! You now have 2 tasks in the list.
+ Bobbing heck! You now have 1 tasks in the list.
 ____________________________________________________________
 ____________________________________________________________
  Aite. I've bobbed it into the list:
    [D][ ] lmao (by:Dec 12 2025 1200)
- Bobbing heck! You now have 3 tasks in the list.
+ Bobbing heck! You now have 2 tasks in the list.
 ____________________________________________________________
 ____________________________________________________________
  Aite. I've bobbed it into the list:
    [E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
- Bobbing heck! You now have 4 tasks in the list.
+ Bobbing heck! You now have 3 tasks in the list.
 ____________________________________________________________
 ____________________________________________________________
  The list do be Bobbing my dude!
  1.[T][ ] hi
- 2.[T][ ] hi
- 3.[D][ ] lmao (by:Dec 12 2025 1200)
- 4.[E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
+ 2.[D][ ] lmao (by:Dec 12 2025 1200)
+ 3.[E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
 ____________________________________________________________
 ____________________________________________________________
  I'm Marking it. I'm Marking it so good!
@@ -42,9 +41,8 @@ ____________________________________________________________
 ____________________________________________________________
  The list do be Bobbing my dude!
  1.[T][ ] hi
- 2.[T][ ] hi
- 3.[D][ ] lmao (by:Dec 12 2025 1200)
- 4.[E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
+ 2.[D][ ] lmao (by:Dec 12 2025 1200)
+ 3.[E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
 ____________________________________________________________
 ____________________________________________________________
  What the Bob!!![BobInvalidFormatException]
@@ -60,8 +58,8 @@ ____________________________________________________________
 ____________________________________________________________
 ____________________________________________________________
   BOB!!! I've removed the task:
-   [T][ ] hi
- Bobbing heck! You now have 3 tasks in the list.
+   [D][ ] lmao (by:Dec 12 2025 1200)
+ Bobbing heck! You now have 2 tasks in the list.
 ____________________________________________________________
 ____________________________________________________________
  What the BOB!!![BobException]
@@ -74,18 +72,17 @@ ____________________________________________________________
 ____________________________________________________________
  The list do be Bobbing my dude!
  1.[T][ ] hi
- 2.[D][ ] lmao (by:Dec 12 2025 1200)
- 3.[E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
-____________________________________________________________
-____________________________________________________________
-  BOB!!! I've removed the task:
-   [D][ ] lmao (by:Dec 12 2025 1200)
- Bobbing heck! You now have 2 tasks in the list.
+ 2.[E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
 ____________________________________________________________
 ____________________________________________________________
   BOB!!! I've removed the task:
    [E][ ] lol (from: Oct 12 2025 1300 to: Oct 12 2025 1400)
  Bobbing heck! You now have 1 tasks in the list.
+____________________________________________________________
+____________________________________________________________
+  BOB!!! I've removed the task:
+   [T][ ] hi
+ Bobbing heck! You now have 0 tasks in the list.
 ____________________________________________________________
 ____________________________________________________________
  Bye have a great time!
